+++ template/src/Theme/Common.js
@@ -3,7 +3,7 @@
  *
  * Use it to define generic component styles (e.g. the default text styles, default button styles...).
  */
-import { Colors } from './_Variables'
+import { Colors } from './Variables'
 import { StyleSheet } from 'react-native'
 
 export default StyleSheet.create({
