+++ template/src/Theme/index.js
@@ -3,4 +3,4 @@ export { default as Gutters } from './Gutters'
 export { default as Images } from './Images'
 export { default as Layout } from './Layout'
 export { default as Common } from './Common'
-export { Colors, FontSize } from './_Variables'
+export { Colors, FontSize } from './Variables'
